requirements.txt
anyio==4.8.0
APScheduler==3.11.0
certifi==2025.1.31
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.0.1
python-telegram-bot==21.10
rsa==4.9
sniffio==1.3.1
Telethon==1.38.1
tzlocal==5.3
